app/backendCopy2/requirements.txt
django>=2.2,<2.3
django-debug-toolbar>=1.7
gunicorn==19.9.0
gevent==1.2.2
psycopg2-binary>=2.7.1,<2.9
whitenoise>=4.1.2
django-crispy-forms>=1.6.1
django-formtools>=2.0
django-settings-export>=1.2.1
djangorestframework>=3.11.2,<3.12
django-rest-multiple-models==2.1.0
requests==2.21.0
minio==7.1.16
coverage>=4.4.2
django-filter>=2.0.0,<2.1
drf-yasg==1.21.7
django-cors-headers==2.2.0
django-extensions==2.0.6
cryptography<=40.0.2
django-reversion==2.0.13
openpyxl==2.5.11
lxml==4.6.3
GDAL==3.0
deepdiff>=4
urllib3>=1.24,<1.25
djangorestframework-csv==2.1.0
geojson==2.4.1
MarkupSafe>=2.0.1
djangorestframework-simplejwt==4.4.0
pyjwt>=2.0,<=2.4.0
thefuzz==0.19.0
geopandas==0.9.0
pyproj==3.0.1
